refactor: [US-043] - Update imports for book_enrichment_service move
