+++ core/management/commands/backfill_covers.py
@@ -6,7 +6,7 @@ import requests
 from django.db.models import Q
 from django.core.management.base import BaseCommand
 
-from core.book_enrichment_service import _clean_title_for_api
+from core.services.book_enrichment_service import _clean_title_for_api
 from core.models import Book
 from core.services._book_urls import cover_url_from_isbn, cover_url_from_olid
